4.21

+++ controllers/blogs.js
@@ -18,9 +18,6 @@ blogsRouter.post('/', async (request, response) => {
   const body = request.body
 
   try {
-    // const token = getTokenFrom(request)
-    // const decodedToken = jwt.verify(token, process.env.SECRET)
-
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
 
     if (!request.token || !decodedToken.id) {
@@ -79,9 +76,20 @@ blogsRouter.get('/:id', async (request, response) => {
 
 blogsRouter.delete('/:id', async (request, response) => {
   try {
-    await Blog.findByIdAndRemove(request.params.id)
-    
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+
+    //Only user that created the blog, can delete it.
+    if(blog.user.toString() === decodedToken.id.toString()) {
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()    
+    } else {
+      response.status(401).send({ error: 'only the creator can delete own blog entries' })
+    }
   } catch (exception) {
     console.log(exception)
     response.status(400).send({ error: 'Unable to delete: malformatted id' })
